+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- **PODAAC-6181**
+  - add relatedUrl subType BROWSE IMAGE SOURCE
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+
 ## [8.6.0]
 ### Added
 - **PODAAC-5876**